request registration
